Refactoring code: Creat a component Folder for home
